Make sure ctname conform to dns-name format requirement
ctname should contains only the characters a-z, 0-9, and - (hyphen).

As ctname is supposed to be extracted from part of the buildroot-generated tarball filename, it will be all fine except when users enable "revision code in filenames", and add their own commits, that will make tarballs with revision code like "r12345+1" in the filenames.

This commit replaces any non-conforming char with "-", and subsequently remove any trailing and leading "-" in ctname.

Note: The OpenWRT buildroot does not generate tarballs with revision number in the filename.
User has to enable this behavior via:
make menuconfig ->" Image Configuration" -> "Version configuration options",   then select
"Version number in filenames" and "Revision code in filenames".

This can help to name multiple lxc instances with their revision codes.