Don't use bind mount for config transfer.
Now the script use the built-in /tmp directory of the ct and pct pull hostside to accomplish config transfer.
Shared bind mount is no longer a requirement, and is only a personal preference for file sharing in the hook script example.
Previously, the OpenWRT ct will become unbootable if the user has not made additional efforts to handle the recreation of the host bind-mount directory upon host boot or before ct start, as the host bind-mount directory will get removed upon every host reboot.
This change can make the initial deployment of the OpenWRT ct using this script simpler.