Add Option for automatic ct hostname generation
If option autoname is set to 0, the script will not generate hostname for the new ct. It will take the hostname from the old ct or the preset ctname instead.
Default value is 1. In fact any value other than 0, the script will always automatically generate hostname for the new ct.