Check if the script invoked by root.
Obviously, root privileges are needed to run this script. Even most users run this in the PVE single user mode, it may not always be the case. Just check it earlier to save some later pct complaints.
Even if this check gets fooled by fakeroot, it is not possible to exec pct anyway.